+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## Next
 
+### Added
+* Ozon Content: Ondersteuning voor `<Inhoud wijzigactie="voegtoe|verwijder">` ([#2528](https://github.com/dso-toolkit/dso-toolkit/issues/2528))
+* Ozon Content: Afbeeldingen in renvooi kunnen weergeven ([#2433](https://github.com/dso-toolkit/dso-toolkit/issues/2433))
+
 ## 📈 62.17.0 - 26-03-2024
 
 ### Added